break function update continue working button problem
