--- android/sdkmanager "platform-tools" "emulator" "
@@ -1,19 +1,0 @@
-sdkmanager "platform-tools" "emulator" "platforms;android-34" "system-images;android-34;google_apis;x86_64"
-
-
-avdmanager create avd -n Pixel_6_Pro_API_34_PlayStore_ARM -k "system-images;android-34;google_apis;x86_64" --device "pixel_6_pro"
-
-sdkmanager --list
-sdkmanager --list_installed
- avdmanager list device "pixel_6_pro"
-
- usr/local/var/lib/android-sdk
-
- avdmanager create avd -h
-
- emulator -avd Pixel_6_Pro_API_34_PlayStore_ARM -gpu on -skin 1440x3120
-
- sdkmanager --uninstall  "system-images;android-34;google_apis_playstore;arm64-v8a"
-
-
- watchman watch-del-all && rm -rf $TMPDIR/react-*  && rm -fr $TMPDIR/metro-cache* && rm -fr $TMPDIR/haste-map-metro*  &&  rm -rf ios/build && xcrun simctl erase all && rm -rf node_modules/ && npm cache verify && npm install && react-native run-ios -- --reset-cache
