+++ package.json
@@ -12,19 +12,19 @@
   "dependencies": {
     "@react-navigation/bottom-tabs": "^6.6.1",
     "@react-navigation/native": "^6.1.18",
-    "@react-navigation/native-stack": "^6.10.1",
+    "@react-navigation/native-stack": "^6.11.0",
     "@react-navigation/stack": "^6.4.1",
     "expo-blur": "^13.0.2",
     "npm": "^10.8.2",
     "pod-install": "^0.2.2",
     "react": "18.2.0",
-    "react-native": "0.74.3",
+    "react-native": "^0.74.3",
     "react-native-camera": "^4.2.1",
     "react-native-elements": "^3.4.3",
     "react-native-fs": "^2.20.0",
     "react-native-paper": "^5.12.5",
     "react-native-safe-area-context": "^4.10.8",
-    "react-native-screens": "^3.32.0",
+    "react-native-screens": "^3.34.0",
     "react-native-share": "^10.2.1",
     "react-native-snap-carousel": "3.9.1",
     "react-native-sqlite-storage": "^6.0.1",
