+++ src/Screens/Break/index.tsx
@@ -1,16 +1,16 @@
 import React from 'react';
 import { View, Text, Button } from 'react-native';
 import BackgroundWrapper from '../../Components/BackgroundWrapper';
-import styles from './defaultCSS'; // Ensure this path is correct
+import styles from './defaultCSS';
 
-const index = ({navigation}) => {
+const Index = ({ navigation }) => {
   return (
     <BackgroundWrapper>
       <View style={styles.container}>
         <Text style={styles.title}>Time for a Break!</Text>
         <Button
           title="Continue Working"
-          onPress={() => navigation.navigate('Welcome')}
+          onPress={() => navigation.navigate('Welcome')} // Navigate to 'Welcome' screen
           color="#4CAF50"
         />
       </View>
@@ -18,4 +18,4 @@ const index = ({navigation}) => {
   );
 };
 
-export default index;
+export default Index;
